abstract base
